use unique function ids
Concatenating function name and accession to create a unique function id.

This is re:#1283

This is working, but I am not commiting to master yet, because when I go through the steps of creating the functional pangenome with the pangenomic tutorial data then something is not right. The problem is most likely in those silly ad-hoc scripts, but I don't want to commit things to master that are not compatible with the tutorial.